requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.9.1
asn1crypto==1.5.1
billiard==4.2.1
blessed==1.21.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.7.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-picklefield==3.3
django-q==1.3.9
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.4.4
openai==1.82.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==3.0.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.0
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-http-client==3.3.7
PyYAML==6.0.2
qrcode==8.2
random-password-generator==2.2.0
redis==3.5.3
reportlab==4.4.1
requests==2.32.4
sendgrid==6.12.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
